+++ fm-dal/src/main/java/org/fm/fm/entity/SysRolePermissionRelation.java
@@ -17,31 +17,31 @@ import java.io.Serializable;
 @EqualsAndHashCode(callSuper = false)
 public class SysRolePermissionRelation extends Model<SysRolePermissionRelation> {
     //角色
-    private Integer roleId;
+    private Long roleId;
     //权限
-    private Integer permissionId;
+    private Long permissionId;
 
     public SysRolePermissionRelation() {
     }
 
-    public SysRolePermissionRelation(Integer roleId, Integer permissionId) {
+    public SysRolePermissionRelation(Long roleId, Long permissionId) {
         this.roleId = roleId;
         this.permissionId = permissionId;
     }
 
-    public Integer getRoleId() {
+    public Long getRoleId() {
         return roleId;
     }
 
-    public void setRoleId(Integer roleId) {
+    public void setRoleId(Long roleId) {
         this.roleId = roleId;
     }
 
-    public Integer getPermissionId() {
+    public Long getPermissionId() {
         return permissionId;
     }
 
-    public void setPermissionId(Integer permissionId) {
+    public void setPermissionId(Long permissionId) {
         this.permissionId = permissionId;
     }
